Small changes to llwindowmacosx: * Don't cast to float, just use doubles for mouse delta * Fix uninitialized values in an NSRect * Removed unused BOOL
Also, fix missing EOF newline

+++ indra/llwindow/llwindowmacosx.h
@@ -123,7 +123,7 @@ public:
 	LLWindowCallbacks* getCallbacks() { return mCallbacks; }
 	LLPreeditor* getPreeditor() { return mPreeditor; }
 	
-	void updateMouseDeltas(float* deltas);
+	void updateMouseDeltas(double* deltas);
 	void getMouseDeltas(float* delta);
 	
 	void handleDragNDrop(std::string url, LLWindowCallbacks::DragNDropAction action);
